requirements.txt
asgiref==3.2.10
autopep8==1.5.4
certifi==2020.6.20
chardet==3.0.4
Django==3.0.8
django-filter==2.3.0
django-test-pep8==0.1
djangorestframework==3.11.0
gunicorn==20.0.4
idna==2.10
importlib-metadata==1.7.0
Markdown==3.2.2
pep8==1.7.1
psycopg2-binary==2.8.5
pycodestyle==2.6.0
pytz==2020.1
requests==2.24.0
sqlparse==0.3.1
toml==0.10.1
urllib3==1.25.10
yapf==0.30.0
zipp==3.1.0
